Create a feature to send env vars/secrets securely Create a retry logic in TelegramAPI for parsing errors Kick off all cron jobs during startup Fix a bug in location request failing for offline process Support different units for temperature and distance Add *args to functions associated to conditions Avoid converting to dict for APIResponse Convert Thread to Timer for reset OTPs Remove revaluation of CronExpression Change headers to use hyphen instead of underscore Add more logging towards unrecognized models Add todos for next release Update .gitignore and docs Release alpha version

+++ jarvis/modules/audio/speaker.py
@@ -132,6 +132,7 @@ def frequently_used(function_name: str) -> NoReturn:
     except Exception as error:  # Commonly raises key error or type error but don't care, log it and remove the file
         logger.error(error)
         logger.warning(data)
-        os.remove(models.fileio.frequent)
+        if os.path.isfile(models.fileio.frequent):
+            os.remove(models.fileio.frequent)
         return
     files.put_frequent(data=data)
